Add Appendix of paper ADB-TRM
This is the appendix section of the paper "Automatic De-Biased Temporal-Relational Modeling for Stock Investment Recommendation", which was accepted by IJCAI 2024 (the 33rd International Joint Conference on Artificial Intelligence). The appendix is divided into two parts: Appendix A and Appendix B. Appendix A provides a detailed description of the search process for the Mutual Information (MI) function mentioned in the paper, and Appendix B elaborates on the adversarial attack aspect discussed in the paper.